Update PlayCanvas and Nextra dependencies, refactor component configurations, and clean up code
