+++ package.json
@@ -28,6 +28,7 @@
         "typescript-eslint": "^8.12.2"
     },
     "dependencies": {
-        "codesandbox": "^2.2.3"
+        "codesandbox": "^2.2.3",
+        "next": "^15.1.7"
     }
 }
