+++ packages/docs/next.config.js
@@ -13,39 +13,39 @@ const withNextra = nextra({
   search: {
     codeblocks: false
   },
-  contentDirBasePath: '/playground',
+  // contentDirBasePath: '/playground',
 })
 
 const nextConfig = {
   reactStrictMode: true,
-  transpilePackages: ['next-mdx-remote-client', 'playcanvas'],
-  async headers() {
-    return [
-        {
-            // matching all API routes
-            source: "/",
-            headers: [
-                { key: "Access-Control-Allow-Credentials", value: "true" },
-                { key: "Access-Control-Allow-Origin", value: "*" }, // replace this your actual origin
-                { key: "Access-Control-Allow-Methods", value: "GET,DELETE,PATCH,POST,PUT" },
-                { key: "Access-Control-Allow-Headers", value: "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version" },
-            ]
-        }
-    ]
-  },
-  redirects: async () => [
-    {
-        source: '/playground',
-        destination: '/playground/model-viewer',
-        permanent: true
-    },
-    {
-      source: '/',
-      destination: '/docs',
-      permanent: true
-    }
-  ]
+  // transpilePackages: ['next-mdx-remote-client', 'playcanvas'],
+  // async headers() {
+  //   return [
+  //       {
+  //           // matching all API routes
+  //           source: "/",
+  //           headers: [
+  //               { key: "Access-Control-Allow-Credentials", value: "true" },
+  //               { key: "Access-Control-Allow-Origin", value: "*" }, // replace this your actual origin
+  //               { key: "Access-Control-Allow-Methods", value: "GET,DELETE,PATCH,POST,PUT" },
+  //               { key: "Access-Control-Allow-Headers", value: "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version" },
+  //           ]
+  //       }
+  //   ]
+  // },
+  // redirects: async () => [
+  //   {
+  //       source: '/playground',
+  //       destination: '/playground/model-viewer',
+  //       permanent: true
+  //   },
+  //   {
+  //     source: '/',
+  //     destination: '/docs',
+  //     permanent: true
+  //   }
+  // ]
 }
 
-export default nextConfig
-// export default withNextra(nextConfig)
+// export default nextConfig
+export default withNextra(nextConfig)
