+++ packages/docs/package.json
@@ -22,8 +22,8 @@
     "motion": "^11.13.1",
     "next": "^15.0.2",
     "next-mdx-remote-client": "^1.0.3",
-    "nextra": "^4.0.0-app-router.42",
-    "nextra-theme-docs": "^4.0.0-app-router.42",
+    "nextra": "^4.2.13",
+    "nextra-theme-docs": "^4.2.13",
     "playcanvas": "^2.3.3",
     "react": "^18.3.1",
     "react-codesandboxer": "^3.1.5",
