+++ packages/docs/src/app/__playground/[[...mdxPath]]/page.tsx
